chore(kurtosis-devnet): split artifact management to its own package
