+++ kurtosis-devnet/pkg/kurtosis/sources/artifact/fs.go
@@ -1,11 +1,10 @@
-package deployer
+package artifact
 
 import (
 	"archive/tar"
 	"bytes"
 	"compress/gzip"
 	"context"
-	"fmt"
 	"io"
 	"path/filepath"
 
@@ -55,11 +54,17 @@ type ArtifactFileWriter struct {
 	writer io.Writer
 }
 
+func NewArtifactFileWriter(path string, writer io.Writer) *ArtifactFileWriter {
+	return &ArtifactFileWriter{
+		path:   path,
+		writer: writer,
+	}
+}
+
 func (a *Artifact) ExtractFiles(writers ...*ArtifactFileWriter) error {
 	paths := make(map[string]io.Writer)
 	for _, writer := range writers {
 		canonicalPath := filepath.Clean(writer.path)
-		canonicalPath = fmt.Sprintf("./%s", canonicalPath)
 		paths[canonicalPath] = writer.writer
 	}
 
@@ -69,11 +74,12 @@ func (a *Artifact) ExtractFiles(writers ...*ArtifactFileWriter) error {
 			break
 		}
 
-		if _, ok := paths[header.Name]; !ok {
+		headerPath := filepath.Clean(header.Name)
+		if _, ok := paths[headerPath]; !ok {
 			continue
 		}
 
-		writer := paths[header.Name]
+		writer := paths[headerPath]
 		_, err = io.Copy(writer, a.reader)
 		if err != nil {
 			return err
